fix unix domain socket

+++ src/Transports/UnixDomainSocketStratergy.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.IO;
 using System.Net.Sockets;
 using System.Threading;
@@ -29,6 +30,7 @@ namespace MediatR.IPC
     {
         private class UnixDomainSocketStratergy : IStreamStratergy<UnixDomainSocketOptions>
         {
+            private static readonly Dictionary<string, Socket> ServerSockets = new();
             private UnixDomainSocketOptions options = new();
 
             async Task<Stream> IStreamStratergy.Provide(StreamType type, string streamName, CancellationToken token)
@@ -38,18 +40,26 @@ namespace MediatR.IPC
                 {
                     var socket = new Socket(AddressFamily.Unix, SocketType.Stream, ProtocolType.Unspecified);
                     token.Register(() => { socket.Close(); });
-                    await Connect(streamName, socket).ConfigureAwait(false);
+                    await Connect(streamName, socket, token).ConfigureAwait(false);
                     return new NetworkStream(socket, FileAccess.ReadWrite, true);
                 }
                 else
                 {
-                    if (File.Exists(streamName))
+                    Socket socket;
+                    lock (ServerSockets)
                     {
-                        File.Delete(streamName);
+                        if (!ServerSockets.TryGetValue(streamName, out socket))
+                        {
+                            if (File.Exists(streamName))
+                            {
+                                // Socket left over from another process.
+                                File.Delete(streamName);
+                            }
+                            socket = ServerSockets[streamName] = new Socket(AddressFamily.Unix, SocketType.Stream, ProtocolType.Unspecified); 
+                            socket.Bind(new UnixDomainSocketEndPoint(streamName));
+                        }
                     }
-
-                    var socket = new Socket(AddressFamily.Unix, SocketType.Stream, ProtocolType.Unspecified);
-                    socket.Bind(new UnixDomainSocketEndPoint(streamName));
+                    
                     socket.Listen(1);
                     Socket? connectedSocket = null;
 #if NET6_0_OR_GREATER
@@ -80,23 +90,27 @@ namespace MediatR.IPC
                         socket.Dispose();
                     }
 #endif
-                    return new NetworkStream(connectedSocket, FileAccess.ReadWrite, true);
+                    return new NetworkStream(connectedSocket, FileAccess.ReadWrite, false);
                 }
             }
 
-            private static async Task Connect(string streamName, Socket socket)
+            private static async Task Connect(string streamName, Socket socket, CancellationToken cancellationToken)
             {
                 bool connected = false;
                 while (!connected)
                 {
                     try
                     {
+#if NET6_0_OR_GREATER
+                        await socket.ConnectAsync(new UnixDomainSocketEndPoint(streamName), cancellationToken).ConfigureAwait(false);
+#else
                         await socket.ConnectAsync(new UnixDomainSocketEndPoint(streamName)).ConfigureAwait(false);
+#endif
                         connected = true;
                     }
                     catch (SocketException)
                     {
-                        await Task.Delay(500).ConfigureAwait(false);
+                        await Task.Delay(500, cancellationToken).ConfigureAwait(false);
                     }
                 }
             }
